HIPT/requirements.txt
astor==0.8.1
autograd==1.4
autograd-gamma==0.5.0
bokeh==2.4.2
braceexpand==0.1.7
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
charset-normalizer==2.1.1
cloudpickle==2.0.0
colorcet==3.0.0
cycler==0.11.0
dask==2022.1.0
datashader==0.13.0
datashape==0.5.2
distributed==2022.1.0
ecos==2.0.10
einops==0.4.1
fonttools==4.29.0
formulaic==0.4.0
fsspec==2022.1.0
future==0.18.2
graphlib-backport==1.0.3
h5py==3.7.0
HeapDict==1.0.1
holoviews==1.14.7
idna==3.3
imageio==2.14.1
interface-meta==1.3.0
joblib==1.1.0
kiwisolver==1.3.2
lifelines==0.27.1
llvmlite==0.38.0
locket==0.2.1
matplotlib==3.5.1
msgpack==1.0.3
multipledispatch==0.6.0
numba==0.55.0
numexpr==2.8.3
numpy==1.20.0
opencv-python==4.4.0.46
osqp==0.6.2.post5
packaging==21.3
pandas==1.4.3
panel==0.12.6
param==1.12.0
partd==1.2.0
Pillow==9.2.0
psutil==5.9.0
pyct==0.4.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
pyviz-comms==2.1.0
PyWavelets==1.2.0
PyYAML==6.0
qdldl==0.1.5.post2
requests==2.28.1
scikit-image==0.19.1
scikit-learn==1.1.2
scikit-survival==0.18.0
scipy==1.9.0
seaborn==0.11.2
six==1.16.0
sortedcontainers==2.4.0
tblib==1.7.0
threadpoolctl==3.1.0
tifffile==2021.11.2
toolz==0.11.2
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.12
webdataset==0.2.18
wrapt==1.14.1
xarray==0.20.2
zict==2.0.0
